+++ src/main/java/de/canitzp/feederhelmet/data/localization/FHLocalizationKeys.java
@@ -5,6 +5,8 @@ public final class FHLocalizationKeys {
     public static final String TAB = "tab.feederhelmet:general";
     public static final String MODULE_FEEDING_DESCRIPTION = "item.feederhelmet:feeder_helmet_module.desc";
     public static final String MODULE_FEEDING_INSTALLED = "item.feederhelmet:feeder_helmet_module_installed.text";
+    public static final String MODULE_FEEDING_REMOVING_DONE = "item.feederhelmet:feeder_helmet_module_removing_done.text";
+    public static final String MODULE_FEEDING_APPLYING_DONE = "item.feederhelmet:feeder_helmet_module_applying_done.text";
     public static final String MODULE_PHOTOSYNTHESIS_DESCRIPTION = "item.feederhelmet:photosynthesis_helmet_module.desc";
     public static final String MODULE_PHOTOSYNTHESIS_INSTALLED = "item.feederhelmet:photosynthesis_module_installed.text";