+++ src/main/java/de/canitzp/feederhelmet/item/ItemFeederModule.java
@@ -2,8 +2,11 @@ package de.canitzp.feederhelmet.item;
 
 import de.canitzp.feederhelmet.data.localization.FHLocalizationKeys;
 import net.minecraft.ChatFormatting;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.Style;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
@@ -15,8 +18,8 @@ import java.util.List;
  */
 public class ItemFeederModule extends Item {
 
-    public ItemFeederModule(){
-        super(new Properties().stacksTo(1));
+    public ItemFeederModule(ResourceLocation resourceLocation){
+        super(new Properties().setId(ResourceKey.create(Registries.ITEM, resourceLocation)).useItemDescriptionPrefix().stacksTo(1));
     }
 
     @Override